pyproject.toml
[project]
name = "telegram-gemini-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiocache==0.12.3",
    "aiodns==3.0.0",
    "aiofiles==23.2.1",
    "aiohttp==3.11.8",
    "beautifulsoup4==4.12.3",
    "cachetools==5.3.3",
    "click==8.1.7",
    "cssselect2==0.8.0",
    "dnspython==2.7.0",
    "fastapi==0.109.2",
    "google-genai==1.10.0",
    "google-generativeai==0.8.3",
    "httpx>=0.27.0",
    "idna==3.3",
    "huggingface-hub==0.30.2",
    "langcodes==3.5.0",
    "lxml==5.3.0",
    "mangum==0.14.0",
    "markdown==3.7",
    "networkx==3.4.2",
    "numpy==2.2.3",
    "opencv-python==4.11.0.86",
    "orjson>=3.8.0",
    "pdfminer-six==20221105",
    "pillow==11.1.0",
    "psutil==7.0.0",
    "pydantic==2.10.2",
    "pydub==0.25.1",
    "pydyf==0.11.0",
    "pymongo==4.9.2",
    "pypdf2==3.0.1",
    "pyphen==0.17.2",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.1.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.18",
    "python-telegram-bot==22.1",
    "regex==2024.11.6",
    "reportlab==4.3.1",
    "requests==2.31.0",
    "safetensors==0.5.3",
    "hf_xet>=0.3.0",
    "scikit-image==0.25.2",
    "pytest==8.4.1",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "faster-whisper>=0.10.0",
    "soundfile>=0.12.1",
    "telegramify-markdown[mermaid]==0.1.17",
    "tenacity==9.0.0",
    "tifffile==2025.2.18",
    "together==1.4.1",
    "tokenizers==0.21.0",
    "tqdm==4.67.1",
    "uvicorn==0.24.0.post1",
    "weasyprint==65.1",
    "msgspec>=0.19.0",
    "fastapi-msgspec>=0.1.0",
    "docxtpl>=0.20.0",
    "spire-doc>=13.4.6",
    "ruff>=0.12.7",
]
